Loading furniture objects through replaceRoomPlaceholder
